requirements.txt
APScheduler==3.8.1
arabic-reshaper==2.1.3
asgiref==3.4.1
auditlog3==1.0.1
backports.zoneinfo==0.2.1
beautifulsoup4==4.10.0
certifi==2021.10.8
charset-normalizer==2.0.7
click==8.0.3
colorama==0.4.4
crypto==1.4.1
cv2-tools==2.4.0
defusedxml==0.7.1
diff-match-patch==20200713
Django==3.2.8
django-bootstrap-datepicker-plus==3.0.5
django-colorfield==0.4.5
django-crispy-forms==1.13.0
django-import-export==2.6.1
django-jet==1.0.8
django-js-asset==1.2.2
django-location-field==2.1.0
django-mptt==0.13.4
django-multiselectfield==0.1.12
django-qr-code==2.2.0
django-smart-selects==1.5.9
django-sslserver==0.22
et-xmlfile==1.1.0
facebook-sdk==3.1.0
filetype==1.0.8
future==0.18.2
geocoder==0.6.0
html5lib==1.1
idna==3.3
ImageHash==4.2.1
jaraco.context==4.0.0
joblib==1.1.0
MarkupPy==1.14
more-itertools==8.10.0
Naked==0.1.31
nltk==3.6.5
numpy==1.21.3
oauthlib==3.1.1
odfpy==1.4.1
opencv-python==4.5.4.58
openpyxl==3.0.9
pandas==1.3.4
patsy==0.5.2
pdfkit==0.6.1
Pillow==8.4.0
psycopg2==2.9.2
PyPDF2==1.26.0
pytesseract==0.3.8
python-bidi==0.4.2
python-constraint==1.4.0
python-dateutil==2.8.2
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
PyWavelets==1.2.0
PyYAML==6.0
pyzbar==0.1.8
razorpay==1.2.0
regex==2021.11.10
reportlab==3.6.2
requests==2.26.0
requests-oauthlib==1.3.0
scikit-learn==1.0.1
scipy==1.7.1
segno==1.3.3
shellescape==3.8.1
six==1.16.0
sklearn==0.0
soupsieve==2.2.1
sqlparse==0.4.2
statsmodels==0.13.0
tablib==3.0.0
textblob==0.17.1
threadpoolctl==3.0.0
torch==1.10.0
torchvision==0.11.1
tqdm==4.62.3
tweepy==3.10.0
typing-extensions==3.10.0.2
tzdata==2021.4
tzlocal==4.0.1
urllib3==1.26.7
webencodings==0.5.1
wikipedia==1.4.0
wolframalpha==5.0.0
xhtml2pdf==0.2.5
xlrd==2.0.1
xlwt==1.3.0
xmltodict==0.12.0